refactor: more moving utils around
