+++ iroh-bytes/src/util.rs
@@ -1,24 +1,15 @@
 //! Utility functions and types.
-use anyhow::{ensure, Context, Result};
-use bao_tree::io::{error::EncodeError, sync::encode_ranges_validated};
-use bytes::Bytes;
-use derive_more::Display;
+use anyhow::{ensure, Result};
 use postcard::experimental::max_size::MaxSize;
-use range_collections::RangeSet2;
 use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
 use std::{
     fmt::{self, Display},
-    io::{Read, Seek, Write},
-    path::{Component, Path, PathBuf},
     result,
     str::FromStr,
 };
 use thiserror::Error;
-use tokio::sync::mpsc;
 pub mod io;
 pub mod progress;
-
-use crate::IROH_BLOCK_SIZE;
 
 /// Hash type used throught.
 #[derive(Debug, PartialEq, Eq, Copy, Clone, Hash)]
@@ -170,182 +161,9 @@ impl From<anyhow::Error> for RpcError {
 #[allow(dead_code)]
 pub type RpcResult<T> = result::Result<T, RpcError>;
 
-/// Todo: gather more information about validation errors. E.g. offset
-///
-/// io::Error should be just the fallback when a more specific error is not available.
-#[derive(Debug, Display, Error)]
-pub enum BaoValidationError {
-    /// Generic io error. We were unable to read the data.
-    IoError(#[from] std::io::Error),
-    /// The data failed to validate
-    EncodeError(#[from] EncodeError),
-}
-
-/// Validate that the data matches the outboard.
-pub fn validate_bao<F: Fn(u64)>(
-    hash: Hash,
-    data_reader: impl Read + Seek,
-    outboard: Bytes,
-    progress: F,
-) -> result::Result<(), BaoValidationError> {
-    let hash = blake3::Hash::from(hash);
-    let outboard =
-        bao_tree::outboard::PreOrderMemOutboardRef::new(hash, IROH_BLOCK_SIZE, &outboard)?;
-
-    // do not wrap the data_reader in a BufReader, that is slow wnen seeking
-    encode_ranges_validated(
-        data_reader,
-        outboard,
-        &RangeSet2::all(),
-        DevNull(0, progress),
-    )?;
-    Ok(())
-}
-
-/// little util that discards data but prints progress every 1MB
-struct DevNull<F>(u64, F);
-
-impl<F: Fn(u64)> Write for DevNull<F> {
-    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        const NOTIFY_EVERY: u64 = 1024 * 1024;
-        let prev = self.0;
-        let curr = prev + buf.len() as u64;
-        if prev % NOTIFY_EVERY != curr % NOTIFY_EVERY {
-            (self.1)(curr);
-        }
-        self.0 = curr;
-        Ok(buf.len())
-    }
-
-    fn flush(&mut self) -> std::io::Result<()> {
-        Ok(())
-    }
-}
-
-/// converts a canonicalized relative path to a string, returning an error if
-/// the path is not valid unicode
-///
-/// this will also fail if the path is non canonical, i.e. contains `..` or `.`,
-/// or if the path components contain any windows or unix path separators
-pub fn canonicalize_path(path: impl AsRef<Path>) -> anyhow::Result<String> {
-    let parts = path
-        .as_ref()
-        .components()
-        .map(|c| {
-            let c = if let Component::Normal(x) = c {
-                x.to_str().context("invalid character in path")?
-            } else {
-                anyhow::bail!("invalid path component {:?}", c)
-            };
-            anyhow::ensure!(
-                !c.contains('/') && !c.contains('\\'),
-                "invalid path component {:?}",
-                c
-            );
-            Ok(c)
-        })
-        .collect::<anyhow::Result<Vec<_>>>()?;
-    Ok(parts.join("/"))
-}
-
-pub(crate) struct ProgressReader<R, F: Fn(ProgressReaderUpdate)> {
-    inner: R,
-    offset: u64,
-    cb: F,
-}
-
-impl<R: Read, F: Fn(ProgressReaderUpdate)> ProgressReader<R, F> {
-    pub fn new(inner: R, cb: F) -> Self {
-        Self {
-            inner,
-            offset: 0,
-            cb,
-        }
-    }
-}
-
-impl<R: Read, F: Fn(ProgressReaderUpdate)> Read for ProgressReader<R, F> {
-    fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
-        let read = self.inner.read(buf)?;
-        self.offset += read as u64;
-        (self.cb)(ProgressReaderUpdate::Progress(self.offset));
-        Ok(read)
-    }
-}
-
-impl<R, F: Fn(ProgressReaderUpdate)> Drop for ProgressReader<R, F> {
-    fn drop(&mut self) {
-        (self.cb)(ProgressReaderUpdate::Done);
-    }
-}
-
-/// Update from a [`ProgressReader`].
-#[derive(Debug, Clone, Copy)]
-pub(crate) enum ProgressReaderUpdate {
-    /// A progress event containing the current offset.
-    Progress(u64),
-    /// The reader has been dropped.
-    Done,
-}
-
-/// A sender for progress messages.
-///
-/// This may optionally be a no-op if the [`Progress::none`] constructor is used.
-#[derive(Debug)]
-pub struct Progress<T>(Option<mpsc::Sender<T>>);
-
-impl<T> Clone for Progress<T> {
-    fn clone(&self) -> Self {
-        Progress(self.0.clone())
-    }
-}
-
-impl<T: fmt::Debug + Send + Sync + 'static> Progress<T> {
-    /// Create a new progress sender.
-    pub fn new(sender: mpsc::Sender<T>) -> Self {
-        Self(Some(sender))
-    }
-
-    /// Create a no-op progress sender.
-    pub fn none() -> Self {
-        Self(None)
-    }
-
-    /// Try to send a message.
-    pub fn try_send(&self, msg: T) {
-        if let Some(progress) = &self.0 {
-            progress.try_send(msg).ok();
-        }
-    }
-
-    /// Block until the message is sent.
-    pub fn blocking_send(&self, msg: T) {
-        if let Some(progress) = &self.0 {
-            progress.blocking_send(msg).ok();
-        }
-    }
-
-    /// Send a message
-    pub async fn send(&self, msg: T) -> anyhow::Result<()> {
-        if let Some(progress) = &self.0 {
-            progress.send(msg).await?;
-        }
-        Ok(())
-    }
-}
-
-/// Create a pathbuf from a name.
-pub fn pathbuf_from_name(name: &str) -> PathBuf {
-    let mut path = PathBuf::new();
-    for part in name.split('/') {
-        path.push(part);
-    }
-    path
-}
-
 /// A non-sendable marker type
 #[derive(Debug)]
-pub struct NonSend {
+pub(crate) struct NonSend {
     _marker: std::marker::PhantomData<std::rc::Rc<()>>,
 }
 
@@ -374,6 +192,6 @@ mod tests {
 
     #[test]
     fn test_canonicalize_path() {
-        assert_eq!(canonicalize_path("foo/bar").unwrap(), "foo/bar");
+        assert_eq!(io::canonicalize_path("foo/bar").unwrap(), "foo/bar");
     }
 }
